chore: 🤖 follow update of ckb change of get_fee_rate_statics

+++ core/src/main/java/org/nervos/ckb/type/FeeRateStatics.java
@@ -4,6 +4,6 @@ package org.nervos.ckb.type;
  * The fee_rate statistics information, includes mean and median
  */
 public class FeeRateStatics {
-    public double mean;
-    public double median;
+    public long mean;
+    public long median;
 }
